+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -3,7 +3,6 @@ export * from './AnthropicCompatibleParameters';
 export * from './AzureParameters';
 export * from './BedrockParameters';
 export * from './CerebrasParameters';
-export * from './GeminiCliParameters';
 export * from './DeepseekParameters';
 export * from './GeminiParameters';
 export * from './GpustackParameters';